Create folders for each auth route
