+++ src/app/auth/login/styles.tsx
@@ -1,0 +1,3 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({});
