+++ src/app/auth/signup/styles.tsx
@@ -1,0 +1,3 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({});
